Refactor device effects pipeline into modular stages
